Advantage bars

+++ src/network/protocol.h
@@ -138,8 +138,8 @@ struct PacketStatusUpdate {
 	
 	uint32_t money[TEAMS_CAP];
 	
-	uint16_t grace_timer;
-	uint16_t time_to_win[TEAMS_CAP];
+	uint32_t grace_timer;
+	uint32_t time_to_win[TEAMS_CAP];
 };
